+++ actions.d.ts
@@ -1,9 +1,5 @@
-import type {ImmutableTree, SomeTreeAction, TreeAction} from './types';
+import type {SomeTreeAction} from './types';
 
 export const NOOP: SomeTreeAction;
 export const REPLACE: SomeTreeAction;
 export const THROW: SomeTreeAction;
-
-export function replaceWith<T>(
-  getValue: (oldValue: T, newValue: T) => T,
-): TreeAction<T>;