+++ actions.mjs
@@ -1,33 +1,15 @@
 // @flow strict
 
 /*::
-import type {ImmutableTree, SomeTreeAction, TreeAction} from './types.mjs';
+import type {SomeTreeAction} from './types.mjs';
 */
 
 export const NOOP/*: SomeTreeAction */ =
-  /*:: <T> */(tree/*: ImmutableTree<T> */)/*: ImmutableTree<T> */ => tree;
+  /*:: <T> */(existingTreeValue/*: T */)/*: T */ => existingTreeValue;
 
 export const REPLACE/*: SomeTreeAction */ =
-  /*:: <T> */(tree/*: ImmutableTree<T> */, value/*: T */)/*: ImmutableTree<T> */ => ({
-    left: tree.left,
-    right: tree.right,
-    size: tree.size,
-    value,
-  });
+  /*:: <T> */(existingTreeValue/*: T */, value/*: T */)/*: T */ => value;
 
 export const THROW/*: SomeTreeAction */ = () => {
   throw new Error('');
 };
-
-export function replaceWith/*:: <T> */(
-  getValue/*: (oldValue: T, newValue: T) => T */,
-)/*: TreeAction<T> */ {
-  return function (tree/*: ImmutableTree<T> */, newValue/*: T */)/*: ImmutableTree<T> */ {
-    return {
-      left: tree.left,
-      right: tree.right,
-      size: tree.size,
-      value: getValue(tree.value, newValue),
-    };
-  };
-}