+++ types.d.ts
@@ -13,7 +13,7 @@ export interface MutableTree<T> {
 }
 
 export type TreeAction<T> =
-  (tree: ImmutableTree<T>, value: T) => ImmutableTree<T>;
+  (existingTreeValue: T, value: T) => T;
 
 export type SomeTreeAction =
-  <T>(tree: ImmutableTree<T>, value: T) => ImmutableTree<T>;
+  <T>(existingTreeValue: T, value: T) => T;